Fix issues with account type retrieval and hiscores' inresponsiveness
The issue was that when a lookup of hiscores for a certain mode (e.g. UIM) failed 3 times, it returned nil and skipped. Because of this, players can be updated with the wrong account type. I fixed this by letting a thread raise the exception to the main thread after 3 failed attempts; the user will then get redirected to their hiscores tab with the notice that the player could not be updated, suggesting to try again.

The check_acc_type was broken since last PR because I accidentally removed the underlying Player.#check_acc_type method. I reointroduced and updated this method again and made sure it handles the before mentioned potential responsiveness issue.

All in all, this should fix the issues that people got when they got assigned a wrong player account type.

I also fixed the rollback/hotfix for checking for HCIM death.